+++ node/middlewares/generateMiddlewares/generateSitemap.ts
@@ -29,11 +29,11 @@ const DEFAULT_REWRITER_ROUTES_PAYLOAD = {
 
 export async function generateSitemap(ctx: EventContext) {
   const { clients: { events }, body: { generationId }, state: { settings } }  = ctx
-  const disableStringRoutes = settings.disableStringRoutes
+  const disableRoutesTerm = settings.disableRoutesTerm
   if (settings.enableNavigationRoutes) {
     events.sendEvent('', GENERATE_REWRITER_ROUTES_EVENT, {
       ...DEFAULT_REWRITER_ROUTES_PAYLOAD,
-      disableStringRoutes,
+      disableRoutesTerm,
       generationId,
     } as RewriterRoutesGenerationEvent)
   }